+++ policyengine-client/src/policyengine/pages/policy/policy.jsx
@@ -50,19 +50,19 @@ export default class Policy extends React.Component {
                 <Col xl={3}>
                     <OverviewHolder>
                         <Affix offsetTop={55}>
-                            <PolicyOverview />
+                            <PolicyOverview page="policy"/>
 				        </Affix>
                         <Affix offsetTop={450}>
                             <SharePolicyLinks page="policy"/>
                             <div className="d-block align-middle">
-                                <div className="d-flex justify-content-center">
+                                <div className="justify-content-center">
                                     {this.context.showPopulationImpact && <NavigationButton 
                                         primary 
                                         target="population-impact" 
                                         text={`Calculate ${this.context.properName} impact`} 
                                     />}
                                 </div>
-                                <div className="d-flex justify-content-center">
+                                <div className="justify-content-center">
                                     {this.context.showHousehold && <NavigationButton
                                         target="household" 
                                         text="Describe your household"